added order history view

+++ Dotnet/Entities/Event.cs
@@ -8,7 +8,7 @@ namespace Shop.Entities
     {
         [Required] [Key] public int Id { get; set; }
         [Required] [StringLength(255)] public string Name { get; set; }
-        [Required] public float Price { get; set; }
+        [Required] public double Price { get; set; }
 
         [Required] [StringLength(255)] public string Place { get; set; }
